+++ pkg/storage/tsdb/caching_bucket.go
@@ -54,10 +54,11 @@ func (cfg *CacheBackend) Validate() error {
 type ChunksCacheConfig struct {
 	CacheBackend `yaml:",inline"`
 
-	SubrangeSize        int64         `yaml:"subrange_size"`
-	MaxGetRangeRequests int           `yaml:"max_get_range_requests"`
-	AttributesTTL       time.Duration `yaml:"attributes_ttl"`
-	SubrangeTTL         time.Duration `yaml:"subrange_ttl"`
+	SubrangeSize               int64         `yaml:"subrange_size"`
+	MaxGetRangeRequests        int           `yaml:"max_get_range_requests"`
+	AttributesTTL              time.Duration `yaml:"attributes_ttl"`
+	AttributesInMemoryMaxItems int           `yaml:"attributes_in_memory_max_items"`
+	SubrangeTTL                time.Duration `yaml:"subrange_ttl"`
 }
 
 func (cfg *ChunksCacheConfig) RegisterFlagsWithPrefix(f *flag.FlagSet, prefix string) {
@@ -68,6 +69,7 @@ func (cfg *ChunksCacheConfig) RegisterFlagsWithPrefix(f *flag.FlagSet, prefix st
 	f.Int64Var(&cfg.SubrangeSize, prefix+"subrange-size", 16000, "Size of each subrange that bucket object is split into for better caching.")
 	f.IntVar(&cfg.MaxGetRangeRequests, prefix+"max-get-range-requests", 3, "Maximum number of sub-GetRange requests that a single GetRange request can be split into when fetching chunks. Zero or negative value = unlimited number of sub-requests.")
 	f.DurationVar(&cfg.AttributesTTL, prefix+"attributes-ttl", 168*time.Hour, "TTL for caching object attributes for chunks. If the metadata cache is configured, attributes will be stored under this cache backend, otherwise attributes are stored in the chunks cache backend.")
+	f.IntVar(&cfg.AttributesInMemoryMaxItems, prefix+"attributes-in-memory-max-items", 0, "Maximum number of object attribute items to keep in a first level in-memory LRU cache. Metadata will be stored and fetched in-memory before hitting the cache backend. 0 to disable the in-memory cache.")
 	f.DurationVar(&cfg.SubrangeTTL, prefix+"subrange-ttl", 24*time.Hour, "TTL for caching individual chunks subranges.")
 }
 
@@ -89,8 +91,6 @@ type MetadataCacheConfig struct {
 	BlockIndexAttributesTTL time.Duration `yaml:"block_index_attributes_ttl"`
 	BucketIndexContentTTL   time.Duration `yaml:"bucket_index_content_ttl"`
 	BucketIndexMaxSize      int           `yaml:"bucket_index_max_size_bytes"`
-	LRUEnabled              bool          `yaml:"lru_enabled"`
-	LRUMaxItems             int           `yaml:"lru_max_items"`
 }
 
 func (cfg *MetadataCacheConfig) RegisterFlagsWithPrefix(f *flag.FlagSet, prefix string) {
@@ -109,8 +109,6 @@ func (cfg *MetadataCacheConfig) RegisterFlagsWithPrefix(f *flag.FlagSet, prefix 
 	f.DurationVar(&cfg.BlockIndexAttributesTTL, prefix+"block-index-attributes-ttl", 168*time.Hour, "How long to cache attributes of the block index.")
 	f.DurationVar(&cfg.BucketIndexContentTTL, prefix+"bucket-index-content-ttl", 5*time.Minute, "How long to cache content of the bucket index.")
 	f.IntVar(&cfg.BucketIndexMaxSize, prefix+"bucket-index-max-size-bytes", 1*1024*1024, "Maximum size of bucket index content to cache in bytes. Caching will be skipped if the content exceeds this size. This is useful to avoid network round trip for large content if the configured caching backend has an hard limit on cached items size (in this case, you should set this limit to the same limit in the caching backend).")
-	f.BoolVar(&cfg.LRUEnabled, prefix+"lru-enabled", false, "Use a first level in memory LRU cache for metadata cache. Metadata will be stored and fetched in memory before hiting the cache backend. **Experimental**")
-	f.IntVar(&cfg.LRUMaxItems, prefix+"lru-max-items", 10000, "Maximum number of items in the first level in memory LRU cache.")
 }
 
 func (cfg *MetadataCacheConfig) Validate() error {
@@ -149,19 +147,22 @@ func CreateCachingBucket(chunksConfig ChunksCacheConfig, metadataConfig Metadata
 	if chunksCache != nil {
 		cachingConfigured = true
 		chunksCache = storecache.NewTracingCache(chunksCache, logger)
-		if metadataCache == nil {
-			metadataCache = chunksCache
-		}
 
-		if metadataConfig.LRUEnabled {
+		// Use the metadata cache for attributes if configured, otherwise fallback to chunks cache.
+		// If in-memory cache is enabled, wrap the attributes cache with the in-memory LRU cache.
+		attributesCache := chunksCache
+		if metadataCache != nil {
+			attributesCache = metadataCache
+		}
+		if chunksConfig.AttributesInMemoryMaxItems > 0 {
 			var err error
-			metadataCache, err = storecache.WrapWithLRUCache(metadataCache, reg, metadataConfig.LRUMaxItems, chunksConfig.AttributesTTL)
+			attributesCache, err = storecache.WrapWithLRUCache(attributesCache, reg, chunksConfig.AttributesInMemoryMaxItems, chunksConfig.AttributesTTL)
 			if err != nil {
 				return nil, errors.Wrapf(err, "lru-metadata-cache")
 			}
 		}
 
-		cfg.CacheGetRange("chunks", chunksCache, isTSDBChunkFile, chunksConfig.SubrangeSize, metadataCache, chunksConfig.AttributesTTL, chunksConfig.SubrangeTTL, chunksConfig.MaxGetRangeRequests)
+		cfg.CacheGetRange("chunks", chunksCache, isTSDBChunkFile, chunksConfig.SubrangeSize, attributesCache, chunksConfig.AttributesTTL, chunksConfig.SubrangeTTL, chunksConfig.MaxGetRangeRequests)
 	}
 
 	if !cachingConfigured {